refactor: package 생성 / 구조 변경
