+++ src/main/java/racingcar/RacingGameMain.java
@@ -1,5 +1,7 @@
 package racingcar;
 
+import racingcar.view.RacingGameRunner;
+
 import java.util.Scanner;
 
 public class RacingGameMain {
