+++ src/main/java/racingcar/view/condition/InputParser.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.view.condition;
 
 import java.util.List;
 import java.util.stream.Collectors;
