+++ src/test/java/racingcar/MoveCarTest.java
@@ -3,6 +3,8 @@ package racingcar;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.RepeatedTest;
 import org.junit.jupiter.api.Test;
+import racingcar.model.Car;
+import racingcar.view.RacingGameRunner;
 
 public class MoveCarTest {
