+++ src/test/java/racingcar/NamingAssignmentTest.java
@@ -4,6 +4,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racingcar.view.condition.InputParser;
 
 import java.util.Arrays;
 import java.util.List;
